requirements-dev.txt
geodesic-api[all]
tesseract-sdk
numpy
earthengine-api